fix: EventAdminAPI 수정
