+++ src/api/EventAdminAPI.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-useless-catch */
 import axios from 'axios';
 
 const accessToken = localStorage.getItem('accessToken');
@@ -16,50 +15,35 @@ export interface EventRequestType {
 }
 
 export const createEvent = async (data: EventRequestType) => {
-  try {
-    const response = await axios.post(`${BASE_URL}/api/v1/admin/events`, data, {
+  const response = await axios.post(`${BASE_URL}/api/v1/admin/events`, data, {
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Authorization_refresh: `Bearer ${refreshToken}`,
+    },
+  });
+  return response;
+};
+
+export const editEvent = async (data: EventRequestType, id: number) => {
+  const response = await axios.patch(
+    `${BASE_URL}/api/v1/admin/events/${id}`,
+    data,
+    {
       headers: {
         Authorization: `Bearer ${accessToken}`,
         Authorization_refresh: `Bearer ${refreshToken}`,
       },
-    });
-    return response;
-  } catch (err) {
-    throw err;
-  }
-};
-
-export const editEvent = async (data: EventRequestType, id: number) => {
-  try {
-    const response = await axios.patch(
-      `${BASE_URL}/api/v1/admin/events/${id}`,
-      data,
-      {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-          Authorization_refresh: `Bearer ${refreshToken}`,
-        },
-      },
-    );
-    return response;
-  } catch (err) {
-    throw err;
-  }
+    },
+  );
+  return response;
 };
 
 export const deleteEvent = async (id: number) => {
-  try {
-    const response = await axios.delete(
-      `${BASE_URL}/api/v1/admin/events/${id}`,
-      {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-          Authorization_refresh: `Bearer ${refreshToken}`,
-        },
-      },
-    );
-    return response;
-  } catch (err) {
-    throw err;
-  }
+  const response = await axios.delete(`${BASE_URL}/api/v1/admin/events/${id}`, {
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      Authorization_refresh: `Bearer ${refreshToken}`,
+    },
+  });
+  return response;
 };
